Changes project and folder names
