+++ deployment/tfhub/tfhub-object-detection/base64/requirements.txt
