+++ deployment/tfhub/tfhub-object-detection/url/.gitignore
