+++ deployment/tfhub/tfhub-object-detection/url/deploy.py
@@ -1,13 +1,9 @@
-import base64
-import io
-import json
-import numpy as np
 import os
 import tempfile
 import tensorflow as tf
 import tensorflow_hub as hub
 import time
-import wget
+import urllib.request
 
 from PIL import Image, ImageOps
 from verta import Client
@@ -24,29 +20,15 @@ class DetectObject(VertaModelBase):
     def __init__(self, artifacts=None):
         module_handle = 'https://tfhub.dev/google/openimages_v4/ssd/mobilenet_v2/1'
         self.detector = hub.load(module_handle).signatures['default']
-
-        url = 'http://s3.amazonaws.com/verta-starter/street-view-images/000001_0.jpg'
-        file = url.split('/')[-1]
-        wget.download(url, file)
-
-        with open(file, 'rb') as img:
-            img_bytes = base64.b64encode(img.read())
-            img_str = img_bytes.decode('utf-8')
-            img_str = json.dumps(img_str)
-            img_str = np.array(img_str).tolist()
-        
-        self.img_base64 = img_str
     
-    def handle_img(self, img, width=640, height=480):
+    def handle_img(self, data, width=640, height=480):
         _, path = tempfile.mkstemp(suffix='.jpg')
-        img_str = json.loads(img)
-        img_bytes = img_str.encode('utf-8')
-        img_bytes = io.BytesIO(base64.b64decode(img_bytes))
-        img_arr = np.array(Image.open(img_bytes), dtype=np.uint8)
-        img = Image.fromarray(img_arr)
+        file_path = f"{tempfile.gettempdir()}/tmp-{data[0].split('/')[-1]}"
+        urllib.request.urlretrieve(data[1], file_path)
+        img = Image.open(file_path)
         img = ImageOps.grayscale(img)
         img.thumbnail((width, height), Image.Resampling.LANCZOS)
-        img.save(path, format = 'JPEG', quality = 90)
+        img.save(path, format='JPEG', quality=90)
         
         print(f"Image downloaded to {path}.")
         return path
@@ -99,10 +81,10 @@ class DetectObject(VertaModelBase):
         
         return result
 
-    def detect_objects(self, file, img_arr):
+    def detect_objects(self, data):
         start_time = time.time()
-        image_path = self.handle_img(img_arr)
-        result = self.run_detector(file, image_path)
+        image_path = self.handle_img(data)
+        result = self.run_detector(data[0], image_path)
         end_time = time.time()
 
         print(f"Inference time: {end_time - start_time}.")
@@ -110,22 +92,21 @@ class DetectObject(VertaModelBase):
 
     @verify_io
     def predict(self, data):
-        file, img_arr = data
-        result = self.detect_objects(file, img_arr)
+        result = self.detect_objects(data)
 
         return result
 
     def describe(self):
-        """Return a description of the service."""        
+        """Return a description of the service."""
         return {
             "method": "predict",
-            "args": f"{self.img_base64}",
+            "args": "http://s3.amazonaws.com/verta-starter/street-view-images/000001_0.jpg",
             "returns": "file, has_car, score, ymin, xmin, ymax, xmax",
             "description": """
-                Identify whether a given object is present in the image, with its score and bounding boxes.
+                Identify whether a given object is present in the URL image, with its score and bounding boxes.
             """,
             "input_description": """
-                A list with base64 images.
+                A list with image URLs.
             """,
             "output_description": """
                 A CSV file with information about all the processed images.
@@ -135,12 +116,12 @@ class DetectObject(VertaModelBase):
     
     def example(self):
         """Return example input json-serializable data."""
-        return ["000001_0.jpg", self.img_base64]
+        return ["000001_0.jpg", "http://s3.amazonaws.com/verta-starter/street-view-images/000001_0.jpg"]
 
 VERTA_HOST = 'app.verta.ai'
-PROJECT_NAME = 'Object Detection V2'
-MODEL_NAME = 'base64'
-ENDPOINT_NAME = 'object-detection-base64'
+PROJECT_NAME = 'Object Detection V1'
+MODEL_NAME = 'url'
+ENDPOINT_NAME = 'object-detection-url'
 
 os.environ['VERTA_EMAIL'] = ''
 os.environ['VERTA_DEV_KEY'] = ''
@@ -168,7 +149,7 @@ output = {
 
 model = registered_model.create_standard_model(
     model_cls = DetectObject,
-    environment = Python(requirements = ['tensorflow', 'tensorflow_hub', 'dill', 'Pillow', 'wget']),
+    environment = Python(requirements = ['tensorflow', 'tensorflow_hub', 'dill', 'Pillow', 'urllib3']),
     model_api = ModelAPI([input], [output]),
     name = MODEL_NAME
 )
@@ -191,13 +172,5 @@ status = endpoint.update(
 assert status['status'] == "active"
 url = 'http://s3.amazonaws.com/verta-starter/street-view-images/000001_0.jpg'
 file = url.split('/')[-1]
-wget.download(url, file)
-
-with open(file, 'rb') as img:
-    img_bytes = base64.b64encode(img.read())
-    img_str = img_bytes.decode('utf-8')
-    img_str = json.dumps(img_str)
-    img_str = np.array(img_str).tolist()
-
-endpoint.get_deployed_model().predict([file, img_str])
+endpoint.get_deployed_model().predict([file, url])
 print(status)
