+++ deployment/tfhub/tfhub-object-detection/url/predict.py
